Add initial scraper implementation and requirements (#1)
* Add initial scraper implementation and requirements

This commit adds the initial implementation of a web scraper in the `_ideas/scraper_try_1` directory. The scraper fetches data from CoinGecko and extracts cryptocurrency names and prices. It uses the `requests` library for making HTTP requests, `BeautifulSoup` for parsing HTML, and `pandas` for data manipulation. The scraper is exposed as a Flask API endpoint at `/api/crypto`. Additionally, the commit includes a `requirements.txt` file with the necessary dependencies for the scraper.

* refactor : removing unnecessary file of req.txt